Added TranspositionTable support
TranspositionTable now uses aging to remove old entries based on the
move number. Mate scores are adjusted when storing and retrieving them
from the table to keep the mate depth accurate.

+++ src/types.hpp
@@ -16,6 +16,6 @@ typedef u32 move_t;
 typedef u64 bitboard_t;
 typedef u16 square_t;
 
-typedef int32_t score_t;
+typedef int16_t score_t;
 
 #define BIT(x) (1L << (x))
